requirements.txt
Django==4.0.1
Pillow==9.0.0
pytesseract==0.3.8

